+++ libvast/src/system/spawn_counter.cpp
@@ -65,8 +65,8 @@ spawn_counter(system::node_actor* self, system::spawn_arguments& args) {
   VAST_ASSERT(index != nullptr);
   VAST_ASSERT(archive != nullptr);
   return self->spawn(counter, std::move(expr), index, archive,
-                     caf::get_or(args.invocation.options,
-                                 "count.skip-candidate-checks", false));
+                     caf::get_or(args.invocation.options, "count.estimate",
+                                 false));
 }
 
 } // namespace vast::system